Modification of the implementation of LambdaStream
